backend/requirements.txt
fastapi
uvicorn
transformers
torch
torchvision
torchaudio
google-generativeai
python-multipart
opencv-python-headless
pillow
numpy
gunicorn